SPI baudrate and send length adjustment
Corrected baudrate SPI init parameter. Shortened length datatype for Spi_send and Uart_send from uint32_t to uint16_t.

+++ inc/buffer.h
@@ -122,7 +122,7 @@ uint16_t Buffer_stat( buffer_param_t *buffer );
 // Adds arbitrary number of bytes to the UART transmission buffer.
 //  Inputs: pointer to a contiguous block of data, the number of bytes
 // Outputs: none
-extern void Uart_send( volatile void* data, uint32_t length );
+extern void Uart_send( volatile void* data, uint16_t length );
 
 #define BUFFER_H_ 1
 #endif
