+++ inc/uart.h
@@ -32,7 +32,7 @@ void Uart_dmaTxHandler( volatile void* data, uint8_t length );
 // Adds arbitrary number of bytes to the UART transmission buffer.
 //  Inputs: pointer to a contiguous block of data, the number of bytes
 // Outputs: none
-void Uart_send( volatile void* data, uint32_t length );
+void Uart_send( volatile void* data, uint16_t length );
 
 #define UART_H_ 1
 #endif
